Refatorando o stream para limitar a quantidade de filmes por ano
